datadog/samples/travel-concierge-agent-observability/concierge_agent/mcp_travel_tools/requirements.txt
boto3>=1.40.8
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp-proto-http>=1.20.0
google-search-results>=2.4.2
mcp>=1.23.3
requests>=2.28.0
